Move TypeRow into own file, introduce TypeRowElem
TypeRowElem =~= Clone + 'static (blanket impl) allows anything in a TypeRow
Move TypeRow(,Elem) into type_row.rs, used by simple.rs but no in-crate deps
PrimType extends TypeRowElem + sealed (+ Debug)
