+++ src/types/simple/serialize.rs
@@ -65,13 +65,6 @@ pub(crate) enum SerSimpleType {
     },
 }
 
-impl super::sealed::Sealed for SerSimpleType {}
-impl PrimType for SerSimpleType {
-    fn tag(&self) -> TypeTag {
-        unimplemented!()
-    }
-}
-
 trait SerializableType: PrimType {
     const TAG: TypeTag;
 }
